autofill ds160/requirements.txt
fastapi
uvicorn
python-docx
pypdf
langchain-openai
langgraph
python-dotenv
annotated-types
